refactor: Organize helper logic
